# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Demo API",
        "description": "Description of your API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/attempts": {
            "get": {
                "tags": [
                    "Attempts"
                ],
                "summary": "List attempts",
                "description": "Get paginated list of attempts with optional filters",
                "operationId": "149a964df8d202f7ffd116053f47a726",
                "parameters": [
                    {
                        "name": "ticket_id",
                        "in": "query",
                        "description": "Filter by ticket ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attempts fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Attempts fetched successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Attempt"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 42
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Attempts"
                ],
                "summary": "Create a new attempt",
                "description": "Creates a new attempt for the authenticated user and assigns random questions from the ticket",
                "operationId": "3c09ee73fe75d0d4b526de0da0b9f625",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ticket_id"
                                ],
                                "properties": {
                                    "ticket_id": {
                                        "description": "ID of the ticket to attempt",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Attempt created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Attempt created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Attempt"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The ticket_id field is required."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error creating attempt: ..."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/attempts/{attempt}": {
            "get": {
                "tags": [
                    "Attempts"
                ],
                "summary": "Get attempt details",
                "description": "Get attempt with ticket, questions, answers and selected answers",
                "operationId": "4ced9d29d15d39c780ef6325cd214b41",
                "parameters": [
                    {
                        "name": "attempt",
                        "in": "path",
                        "description": "Attempt ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attempt fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Attempt created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Attempt"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Attempt not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/attempts/{attempt}/submit": {
            "post": {
                "tags": [
                    "Attempts"
                ],
                "summary": "Submit attempt and calculate score",
                "operationId": "43a8ea70738115ead3da8e6d73891d7e",
                "parameters": [
                    {
                        "name": "attempt",
                        "in": "path",
                        "description": "Attempt ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attempt submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Attempt submitted successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Attempt"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Attempt not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Answer": {
                "title": "Answer",
                "required": [
                    "id",
                    "question_id",
                    "content",
                    "is_correct"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "question_id": {
                        "description": "ID of the related question",
                        "type": "integer",
                        "example": 2
                    },
                    "content": {
                        "description": "Answer content",
                        "type": "string",
                        "example": "42"
                    },
                    "is_correct": {
                        "description": "Indicates if this is the correct answer",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "Attempt": {
                "title": "Attempt",
                "required": [
                    "id",
                    "ticket_id",
                    "user_id",
                    "score",
                    "started_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "ticket_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 7
                    },
                    "score": {
                        "type": "integer",
                        "example": 0
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-21T10:15:00Z"
                    },
                    "finished_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    },
                    "ticket": {
                        "$ref": "#/components/schemas/Ticket"
                    },
                    "attempt_answers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AttemptAnswer"
                        }
                    }
                },
                "type": "object"
            },
            "AttemptAnswer": {
                "title": "AttemptAnswer",
                "required": [
                    "id",
                    "attempt_id",
                    "question_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "attempt_id": {
                        "description": "ID of the related attempt",
                        "type": "integer",
                        "example": 5
                    },
                    "question_id": {
                        "description": "ID of the related question",
                        "type": "integer",
                        "example": 2
                    },
                    "answer_id": {
                        "description": "ID of the selected answer, nullable if not answered yet",
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "question": {
                        "$ref": "#/components/schemas/Question"
                    },
                    "answer": {
                        "$ref": "#/components/schemas/Answer"
                    }
                },
                "type": "object"
            },
            "Question": {
                "title": "Question",
                "required": [
                    "id",
                    "ticket_id",
                    "content"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "ticket_id": {
                        "description": "ID of the related ticket",
                        "type": "integer",
                        "example": 3
                    },
                    "content": {
                        "description": "Question text",
                        "type": "string",
                        "example": "What is 2 + 2?"
                    },
                    "image_url": {
                        "description": "Optional image URL for the question",
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "answers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Answer"
                        }
                    }
                },
                "type": "object"
            },
            "Ticket": {
                "title": "Ticket",
                "required": [
                    "id",
                    "title",
                    "is_active"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Ticket title",
                        "type": "string",
                        "example": "Math Quiz"
                    },
                    "description": {
                        "description": "Optional description of the ticket",
                        "type": "string",
                        "example": "A set of math questions",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Indicates if the ticket is active",
                        "type": "boolean",
                        "example": true
                    },
                    "questions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Question"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}